Address pageable promotion review
Clarify the explicit extension remediation and share collection-path detection between ARM lint rules.
