+++ packages/typespec-azure-resource-manager/src/rules/list-operation-missing-pageable.md
@@ -2,9 +2,9 @@ ARM collection GET operations must use pageable response metadata. Authoring pag
 keeps generated SDK list methods consistent and avoids a later breaking change when a service needs
 to add continuation links.
 
-Use the standard ARM list operation templates when possible. Custom list operations must use
+Use the standard ARM list operation templates when possible. Custom list operations can use
 `@list`, identify the returned items with `@pageItems`, and identify the continuation link with
-`@nextLink`.
+`@nextLink`. Alternatively, they can explicitly author a truthy `x-ms-pageable` OpenAPI extension.
 
 ## ❌ Incorrect
